# perf(init): Use random for session-key instead of invoking starship session