@matchai 2020-10-23
Testing this change locally, I seem to be seeing the inverse perf outcome on my own machine:

```
starship on  master [$] is  v0.46.2 via  v1.47.0
❯ hyperfine "starship prompt" -m 100
Benchmark #1: starship prompt
  Time (mean ± σ):      47.1 ms ±   2.9 ms    [User: 28.6 ms, System: 46.3 ms]
  Range (min … max):    43.8 ms …  72.8 ms    100 runs

starship on  master [$] is  v0.46.2 via  v1.47.0
❯ hub pr checkout 1755
Switched to branch 'use-random-for-session-id'

starship on  use-random-for-session-id [$] is  v0.45.2 via  v1.47.0
❯ cargo install --path=. --force
  Installing starship v0.45.2 (/Users/matan/dev/starship)
    Updating crates.io index
   Compiling starship v0.45.2 (/Users/matan/dev/starship)
    Finished release [optimized] target(s) in 3m 08s
   Replacing /Users/matan/.cargo/bin/starship
    Replaced package `starship v0.46.2 (/Users/matan/dev/starship)` with `starship v0.45.2 (/Users/matan/dev/starship)` (executable `starship`)

starship on  use-random-for-session-id [$] is  v0.45.2 via  v1.47.0 took 3m8s
❯ hyperfine "starship prompt" -m 100
Benchmark #1: starship prompt
  Time (mean ± σ):      78.7 ms ±   3.0 ms    [User: 46.2 ms, System: 92.6 ms]
  Range (min … max):    59.0 ms …  84.2 ms    100 runs
```

I'd be curious to see others try this change locally to see if I'm an outlier.
I'm running zsh 5.8 (x86_64-apple-darwin19.3.0).