@Gautham 2020-10-23
Ok. That's odd (both because it's different from my experience and because it shouldn't happen theoretically 😅)

Would you mind setting the session key to a constant value and checking the performance for that as well?