@Gautham 2020-10-23
Note: The machines themselves have different config so only compare no-starship vs normal-starship vs starship-with-random within the context of an env.

Based on the below, I can definitely see a drop in latency on MBP with less measurable (but not non-existent) improvement in EC2 as well.

```
[EC2 Machine] ZSH: 5.8

No Starship
---------------------------
Benchmark #1: zsh -i -c exit;
  Time (mean ± σ):      29.3 ms ±   1.0 ms    [User: 9.4 ms, System: 2.0 ms]
  Range (min … max):    28.5 ms …  33.7 ms    50 runs

Default Starship
---------------------------
Benchmark #1: zsh -i -c exit;
  Time (mean ± σ):      33.5 ms ±   1.2 ms    [User: 13.8 ms, System: 6.2 ms]
  Range (min … max):    31.7 ms …  39.0 ms    50 runs

Starship Using $RANDOM
---------------------------
Benchmark #1: zsh -i -c exit;
  Time (mean ± σ):      30.9 ms ±   4.6 ms    [User: 11.0 ms, System: 3.0 ms]
  Range (min … max):    28.7 ms …  56.9 ms    50 runs
  
[Macbook Pro] ZSH: zsh 5.8 (x86_64-apple-darwin19.3.0)

No Starship
---------------------------
Benchmark #1: zsh -i -c exit;
  Time (mean ± σ):      60.8 ms ±   3.7 ms    [User: 30.5 ms, System: 23.8 ms]
  Range (min … max):    57.2 ms …  74.8 ms    50 runs

Default Starship
---------------------------
Benchmark #1: zsh -i -c exit;
  Time (mean ± σ):      75.4 ms ±   5.6 ms    [User: 36.5 ms, System: 31.0 ms]
  Range (min … max):    66.8 ms …  86.0 ms    50 runs

Starship Using $RANDOM
---------------------------
Benchmark #1: zsh -i -c exit;
  Time (mean ± σ):      61.5 ms ±   3.1 ms    [User: 30.8 ms, System: 24.5 ms]
  Range (min … max):    58.3 ms …  76.3 ms    50 runs


[Macbook Pro] Bash - 5.0.18(1)-release

No Starship
---------------------------
Benchmark #1: bash -i -c exit;
  Time (mean ± σ):       5.8 ms ±   0.4 ms    [User: 1.4 ms, System: 2.3 ms]
  Range (min … max):     5.2 ms …   7.0 ms    50 runs

Default Starship
---------------------------
Benchmark #1: bash -i -c exit;
  Time (mean ± σ):      30.7 ms ±   1.7 ms    [User: 11.0 ms, System: 15.0 ms]
  Range (min … max):    26.2 ms …  35.3 ms    50 runs

Starship Using $RANDOM
---------------------------
Benchmark #1: bash -i -c exit;
  Time (mean ± σ):      22.1 ms ±   1.9 ms    [User: 8.0 ms, System: 11.7 ms]
  Range (min … max):    17.9 ms …  26.4 ms    50 runs


[Macbook Pro] 3.2.57(1)-release

No Starship
---------------------------
Benchmark #1: /bin/bash -i -c exit;
  Time (mean ± σ):       6.2 ms ±   0.4 ms    [User: 1.1 ms, System: 2.8 ms]
  Range (min … max):     5.4 ms …   7.7 ms    50 runs

Default Starship
---------------------------
Benchmark #1: /bin/bash -i -c exit;
  Time (mean ± σ):      29.4 ms ±   1.0 ms    [User: 10.2 ms, System: 15.1 ms]
  Range (min … max):    28.0 ms …  32.5 ms    50 runs

Starship Using $RANDOM
---------------------------
Benchmark #1: /bin/bash -i -c exit;
  Time (mean ± σ):      22.9 ms ±   2.3 ms    [User: 7.5 ms, System: 13.5 ms]
  Range (min … max):    20.2 ms …  31.1 ms    50 runs
```