@Gautham 2020-10-23
> Testing this change locally, I seem to be seeing the inverse perf outcome on my own machine:
> 
> ```
> ### `master` installed
> ❯ hyperfine "starship prompt" -m 100
> Benchmark #1: starship prompt
>   Time (mean ± σ):      47.1 ms ±   2.9 ms    [User: 28.6 ms, System: 46.3 ms]
>   Range (min … max):    43.8 ms …  72.8 ms    100 runs
> 
> 
> ### `pr 1755` installed
> ❯ hyperfine "starship prompt" -m 100
> Benchmark #1: starship prompt
>   Time (mean ± σ):      78.7 ms ±   3.0 ms    [User: 46.2 ms, System: 92.6 ms]
>   Range (min … max):    59.0 ms …  84.2 ms    100 runs
> ```
> 
> I'd be curious to see others try this change locally to see if I'm an outlier.
> I'm running zsh 5.8 (x86_64-apple-darwin19.3.0).

Ok, I just realized two things -

1. You are benchmarking `starship prompt`, which is the time it takes to print the prompt once starship is invoked. My fix is in the wrapper that calls starship from zsh/bash. You'd want to do `hyperfine -r 50 "zsh -i -c exit;"` to benchmark the perf diff here.
1. The reason for the anomaly above (as we discussed on Discord) is that my PR is 2w old and behind master so missing a couple of perf-improving commits. Now that's rebased to HEAD, you should see less anomalies.