@davidkna 2020-11-03
@andytom Powershell could use [this](https://github.com/ATiltedTree/starship/blob/b55f642305498921d4dc0fb183fb209fb9040eb5/src/init/starship.ps1#L29)