@Gautham 2020-11-06
I limited this to zsh and bash because those are the shells I'm familiar with and backward-compatibility is important so we need to be careful to check any feature used would be available for majority of the userbase. I can probably look into fish and ion as a separate PR when I get time though when time permits. I don't have a Windows laptop handy anymore so I can't help much with Powershell.

On next steps: Can this be merged as-is or do I need to rebase it first? (Or is anything else pending for this?)
PS: Thanks for the reviews!