@davidkna 2020-11-06
I think this could have likely been merged as is, but it looks like you already rebased this PR.
BTW powershell actually also [runs on macos and linux](https://github.com/PowerShell/PowerShell/releases/latest). But I've just tested the snippet I linked above in the older Powershell 5 that comes with windows instead of powershell 7 and it failed to run, so it might be better to handle this in a different PR.