@Gautham 2020-11-06
Thanks for confirming. I somehow missed the fact that PowerShell was now cross-platform -_-

Btw, it looks like the test that failed in the workflow is just a flaky one and not related to my changes.